Updates to various state APIs
The vehicle_state API has changed a lot, with a lot of the
elements moving to a new API (not yet implemented).

The new seat heater results are not currently correct (at least not
on my car), but the value of "false" that is now returned by the API was
causing crashes. Will investigate this more as I have time.